Modularise .js files
